Move vespa search result order tests to separate file
