+++ tests/search/vespa/test_vespasearch.py
@@ -435,50 +435,6 @@ def test_multiple_filters(label, query, test_vespa, data_db, monkeypatch, data_c
 
 
 @pytest.mark.search
-@pytest.mark.parametrize("label, query", [("search", "the"), ("browse", "")])
-def test_result_order_score(
-    label, query, test_vespa, data_db, monkeypatch, data_client
-):
-    monkeypatch.setattr(search, "_VESPA_CONNECTION", test_vespa)
-    _populate_db_families(data_db)
-
-    params = {
-        "query_string": query,
-        "sort_field": "date",
-        "sort_order": "asc",
-    }
-    asc_date_body = _make_search_request(data_client, params)
-    asc_dates = [f["family_date"] for f in asc_date_body["families"]]
-
-    params["sort_order"] = "desc"
-    desc_date_body = _make_search_request(data_client, params)
-    desc_dates = [f["family_date"] for f in desc_date_body["families"]]
-
-    assert VESPA_FIXTURE_COUNT == len(asc_dates) == len(desc_dates)
-    assert asc_dates == list(reversed(desc_dates))
-    assert asc_dates[0] < desc_dates[0]
-    assert asc_dates[-1] > desc_dates[-1]
-
-
-@pytest.mark.search
-@pytest.mark.parametrize("label, query", [("search", "the"), ("browse", "")])
-def test_result_order_title(
-    label, query, test_vespa, data_db, monkeypatch, data_client
-):
-    monkeypatch.setattr(search, "_VESPA_CONNECTION", test_vespa)
-    _populate_db_families(data_db)
-
-    params = {
-        "query_string": query,
-        "sort_field": "title",
-        "sort_order": "asc",
-    }
-
-    # Scope of test is to confirm this does not cause a failure
-    _ = _make_search_request(data_client, params)
-
-
-@pytest.mark.search
 def test_continuation_token__families(test_vespa, data_db, monkeypatch, data_client):
     monkeypatch.setattr(search, "_VESPA_CONNECTION", test_vespa)
